+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Fix: map was giving inconsistent results 0.6.0.dev8 #116
 * Fix: tui continuously rebuilds if an input file exists in a parent directory
   to the output directory 0.6.0.dev9 #118
+* Feat: report the cache stats for the current run 0.6.0.dev11 #121
 
 ## 0.5.2